Moved InferenceService class to another file
